requirements.txt
pyspark==3.1.1
PyYAML==5.4.1
